requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
redis>=5.0.0
gunicorn>=21.0.0
httpx>=0.27.0
google-generativeai>=0.3.2
mcp>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
PyYAML>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0 
psutil>=5.9.0 

# Google Cloud dependencies
google-cloud-logging>=3.0.0
google-cloud-redis>=2.0.0
google-cloud-run>=0.10.0
